requirements.txt
anyio==3.6.2
bidict==0.22.1
fastapi==0.93.0
idna==3.4
pydantic==1.10.6
python-engineio==4.3.4
python-socketio==5.7.2
sniffio==1.3.0
starlette==0.25.0
typing_extensions==4.5.0
